backend/src/requirements.txt
odfpy
fastapi
uvicorn